requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
motor>=3.1.2
pymongo>=4.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0